🔄 synced local 'skyvern/' with remote 'skyvern/'
<!-- ELLIPSIS_HIDDEN -->

> [!IMPORTANT]
> Migrate block and workflow related schemas to `skyvern/schemas/workflows.py` and update imports across the codebase.
>
>   - **Schema Migration**:
>     - Migrate `BlockType`, `BlockStatus`, `WorkflowStatus`, `BlockResult`, `FileType`, `FileStorageType`, and YAML classes from `skyvern/forge/sdk/workflow/models/block.py`, `constants.py`, and `yaml.py` to `skyvern/schemas/workflows.py`.
>     - Remove `constants.py` and `yaml.py` as they are no longer needed.
>   - **Import Updates**:
>     - Update imports in `agent_functions.py`, `migrate_workflows_to_workflow_run_blocks.py`, `transform_workflow_run.py`, `test_helpers.py`, `client.py`, `utils.py`, `agent_protocol.py`, `run_blocks.py`, `workflow_runs.py`, `workflow.py`, `service.py`, and `task_v2_service.py` to use new schema paths.
>   - **Code Cleanup**:
>     - Remove redundant class definitions from `block.py` and `workflow.py` after migration.
>     - Adjust related logic in `service.py` and `task_v2_service.py` to accommodate schema changes.
>
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for a104c34dbf264ef0b33083f439e4ed5e5a4e4e90. You can [customize](https://app.ellipsis.dev/Skyvern-AI/settings/summaries) this summary. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->
