+++ skyvern/forge/sdk/schemas/workflow_runs.py
@@ -7,8 +7,8 @@ from typing import Any
 from pydantic import BaseModel
 
 from skyvern.forge.sdk.schemas.task_v2 import Thought
-from skyvern.forge.sdk.workflow.models.block import BlockType
 from skyvern.schemas.runs import RunEngine
+from skyvern.schemas.workflows import BlockType
 from skyvern.webeye.actions.actions import Action
