--- skyvern/forge/sdk/workflow/models/constants.py
@@ -1,6 +1,0 @@
-from enum import StrEnum
-
-
-class FileStorageType(StrEnum):
-    S3 = "s3"
-    AZURE = "azure"
